pyproject.toml
[tool.poetry]
authors = ["Leon Bowie <leon@bowie-co.nz>"]
description = ""
license = "AGPLv3"
name = "app"
version = "1.0.0"

[tool.poetry.dependencies]
Pillow = "^8.0.1"
SQLAlchemy = "^1.3.20"
aiofiles = "^0.6.0"
aiohttp = "^3.7.2"
alembic = "^1.4.3"
asgi-ratelimit = {extras = ["full"], version = "^0.4.0"}
asyncpg = "^0.21.0"
databases = "^0.4.0"
fastapi = "^0.61.1"
gunicorn = "^20.0.4"
loguru = "^0.5.3"
minecraft-data = "^2.70.1"
opencv-python = "^4.4.0"
pydantic = "^1.7.1"
python = "^3.7"
python-dotenv = "^0.15.0"
sentry-sdk = "^0.19.3"
tqdm = "^4.54.1"
uvicorn = "^0.12.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = {extras = ["toml"], version = "^5.3"}
darglint = "^1.5.5"
flake8 = "^3.8.4"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^20.1.4"
flake8-docstrings = "^1.5.0"
mypy = "^0.790"
nox = "^2020.8.22"
nox-poetry = "^0.5.0"
pep8-naming = "^0.11.1"
pre-commit = "^2.8.2"
pre-commit-hooks = "^3.3.0"
pytest = "^6.1.2"
pytest-asyncio = "^0.14.0"
pytest-cov = "^2.10.1"
reorder-python-imports = "^2.3.6"
safety = "^1.9.0"
typeguard = "^2.10.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.1.0"]
